+++ src/library/extendedKalmanFilter.cpp
@@ -32,13 +32,29 @@
 
 #define TsP                       10e-3  /* Position control sampling time */
 
-
-
 namespace teamsannio_med_control {
 
-ExtendedKalmanFilter::ExtendedKalmanFilter() {
+ExtendedKalmanFilter::ExtendedKalmanFilter()
+       :Xp_(Eigen::VectorXf::Zero(6)), 
+        Xe_(Eigen::VectorXf::Zero(6)),
+	P_(Eigen::MatrixXf::Zero(6,6)),
+        Pe_(Eigen::MatrixXf::Zero(6,6)),
+        Hatx_(Eigen::VectorXf::Zero(6)),
+        u_T_private_(0),
+        Hp_(Eigen::MatrixXf::Identity(6,6)){
+
+              	
+		A_private_ <<   1, 0, 0,  TsP,    0,    0,
+				0, 1, 0,     0, TsP,    0,
+				0, 0, 1, 	   0,    0, TsP,
+				0, 0, 0,     1,    0,    0,
+				0, 0, 0,     0,    1,    0,
+				0, 0, 0,     0,    0,    1;
+
+		Qp_std_ = Qp_private_.transpose()*Qp_private_;
+                 	
+                Rp_std_ = Rp_private_.transpose()*Rp_private_;
 	
-	timer11_ = n11_.createTimer(ros::Duration(TsP), &ExtendedKalmanFilter::CallbackEstimate, this, false, true); 
 	
 }
 
@@ -49,38 +65,29 @@ void ExtendedKalmanFilter::SetOdometry(const EigenOdometry& odometry) {
     odometry_private_ = odometry;    
 }
 
+void ExtendedKalmanFilter::SetFilterParameters(FilterParameters *filter_parameters_){
+
+     Rp_private_ = filter_parameters_->Rp_; 
+     Qp_private_ = filter_parameters_->Qp_;
+}
+
 void ExtendedKalmanFilter::Estimator(state_t *state_, EigenOdometry* odometry_){
    assert(state_);
    assert(odometry_);
 
    SetOdometry(*odometry_);
 
-   state_->position.x = odometry_private_.position[0];
-   state_->position.y = odometry_private_.position[1];
-   state_->position.z = odometry_private_.position[2];
-  
-   state_->linearVelocity.x = odometry_private_.velocity[0];
-   state_->linearVelocity.y = odometry_private_.velocity[1];
-   state_->linearVelocity.z = odometry_private_.velocity[2];
+   Predict();
+   Correct();
 
-   state_->attitude.roll = roll;
-   state_->attitude.pitch = pitch;
-   state_->attitude.yaw = yaw;
+   state_->position.x = Xe_(0);
+   state_->position.y = Xe_(1);
+   state_->position.z = Xe_(2);
+
+   state_->linearVelocity.x = Xe_(3);
+   state_->linearVelocity.x = Xe_(4);
+   state_->linearVelocity.x = Xe_(5);   
  
-   state_->angularVelocity.x = odometry_private_.angular_velocity[0];
-   state_->angularVelocity.y = odometry_private_.angular_velocity[1];
-   state_->angularVelocity.z = odometry_private_.angular_velocity[2];
- 
-}
-
-void ExtendedKalmanFilter::SetVehicleParameters(){
-
-      m_ = vehicle_parameters_.mass_;
-      g_ = vehicle_parameters_.gravity_;
-      Ix_ = vehicle_parameters_.inertia_(0,0);
-      Iy_ = vehicle_parameters_.inertia_(1,1);
-      Iz_ = vehicle_parameters_.inertia_(2,2);
-
 }
 
 void ExtendedKalmanFilter::Quaternion2Euler(double* roll, double* pitch, double* yaw) const {
@@ -100,121 +107,87 @@ void ExtendedKalmanFilter::Quaternion2Euler(double* roll, double* pitch, double*
 	
 }
 
-void ExtendedKalmanFilter::predict(Eigen::Vector6f* xp, Eigen::Matrix6f* P){
-    assert(xp);
-    assert(P);
+
+void ExtendedKalmanFilter::SetThrustCommand(double u_T){
+
+    u_T_private_ = u_T;
+
+}
+
+void ExtendedKalmanFilter::SetVehicleParameters(double m, double g){
+
+      m_private_ = m;
+      g_private_ = g;
+      
+}
+
+void ExtendedKalmanFilter::Predict(){
+
+    Quaternion2Euler(&phi_, &theta_, &psi_);
 
     double x, y, z, dx, dy, dz;
-    x  = odometry_private_.position[0];
-    y  = odometry_private_.position[1];
-    z  = odometry_private_.position[2];
-	dx = odometry_private_.velocity[0];
-	dy = odometry_private_.velocity[1];
-	dz = odometry_private_.velocity[2];
-	
-	phi = state.attitude.roll;
-	theta = state.attitude.pitch;
-	psi = state.attitude.yaw;
-	
-	
-	dot_x = (cos(theta) * cos(psi) * dx) + 
-           ( ( (sin(phi) * sin(theta) * cos(psi) ) - ( cos(phi) * sin(psi) ) ) * dy) + 
-           ( ( (cos(phi) * sin(theta) * cos(psi) ) + ( sin(phi) * sin(psi) ) ) * dz); 
+    x  = Hatx_(0);
+    y  = Hatx_(1);
+    z  = Hatx_(2);
 
-    dot_y = (cos(theta) * sin(psi) * dx) +
-           ( ( (sin(phi) * sin(theta) * sin(psi) ) + ( cos(phi) * cos(psi) ) ) * dy) +
-           ( ( (cos(phi) * sin(theta) * sin(psi) ) - ( sin(phi) * cos(psi) ) ) * dz);
+    dx = Hatx_(3);
+    dy = Hatx_(4);
+    dz = Hatx_(5);
 
-   dot_z = (-sin(theta) * dx) + ( sin(phi) * cos(theta) * dy) +
-           (cos(phi) * cos(theta) * dz);
-   
-   
-   double u_x, u_y, u_T, u_Terr;
-   PositionController::PosController(&u_x, &u_y, &u_T, &u_Terr);
+    double dx_ENU, dy_ENU, dz_ENU;
 	
-	// Non linear Position model discretized with forward Euler
-	x = x + Tsp_ * dot_x;
-	y = y + Tsp_ * dot_y;
-	z = z + Tsp_ * dot_z;
-	dot_x = dot_x + Tsp_ * (*u_T/m_ * (cos(psi) * sin(theta) * cos(phi) + sin(psi) * sin(theta)));
-	dot_y = dot_y + Tsp_ * (*u_T/m_ * (sin(psi) * sin(theta) * cos(phi) - cos(psi) * sin(phi)));
-	dot_z = dot_z + Tsp_ * (-g_ + *u_T/m_ * (cos(theta) * cos(phi));
+    dx_ENU = (cos(theta_) * cos(psi_) * dx) + 
+	     ( ( (sin(phi_) * sin(theta_) * cos(psi_) ) - ( cos(phi_) * sin(psi_) ) ) * dy) + 
+	     ( ( (cos(phi_) * sin(theta_) * cos(psi_) ) + ( sin(phi_) * sin(psi_) ) ) * dz); 
+
+    dy_ENU = (cos(theta_) * sin(psi_) * dx) +
+	     ( ( (sin(phi_) * sin(theta_) * sin(psi_) ) + ( cos(phi_) * cos(psi_) ) ) * dy) +
+	     ( ( (cos(phi_) * sin(theta_) * sin(psi_) ) - ( sin(phi_) * cos(psi_) ) ) * dz);
+
+    dz_ENU = (-sin(theta_) * dx) + ( sin(phi_) * cos(theta_) * dy) +
+	     (cos(phi_) * cos(theta_) * dz);
+
+
+     //Nonlinear state propagation 
+     x = x + TsP * dx_ENU;
+     y = y + TsP * dy_ENU;
+     z = z + TsP * dz_ENU;
+     dx_ENU = dx_ENU + TsP * (u_T_private_ * (1/m_private_) * (cos(psi_) * sin(theta_) * cos(phi_) + sin(psi_) * sin(theta_)));
+     dy_ENU = dy_ENU + TsP * (u_T_private_ * (1/m_private_) * (sin(psi_) * sin(theta_) * cos(phi_) - cos(psi_) * sin(phi_)));
+     dz_ENU = dz_ENU + TsP * (-g_private_ + u_T_private_ * (1/m_private_) * (cos(theta_) * cos(phi_)));
 	
-	// Jacobian Matrix
-	Eigen::Matrix6f A;
-	
-	A <<  1, 0, 0,  Tsp_,    0,    0,
-		  0, 1, 0,     0, Tsp_,    0,
-		  0, 0, 1, 	   0,    0, Tsp_,
-		  0, 0, 0,     1,    0,    0,
-		  0, 0, 0,     0,    1,    0,
-		  0, 0, 0,     0,    0,    1;
-		  
-	Eigen::Matrix6f Qp;
-	
-	Qp <<     1e-6,			0,			0,			0,			0,			0,
-				 0,	     1e-6,			0,			0,			0,			0,
-				 0,			0,       1e-6,			0,			0,			0,
-				 0,			0,			0,       1e-6,			0,			0,
-				 0,			0,			0,			0,       1e-6, 			0,
-				 0,			0,			0,			0,			0,	     1e-6;
-				 
-	Qp = pow(Qp,2);
-	
-	// Prediction error Matrix
-	Eigen::MatrixXd P = A * P * A.transpose() + Qp;
-	
-	Eigen::VectorXd xp(6);
-	
-	xp(1) = x;
-	xp(2) = y;
-	xp(3) = z;
-	xp(4) = dot_x;
-	xp(5) = dot_y;
-	xp(6) = dot_z;
-				 
+		 				 
+     // Prediction error Matrix
+     P_ = A_private_*(P_)*A_private_.transpose() + Qp_std_;
+		
+     //The predicted state
+     Xp_ << x, y, z, dx, dy, dz;
+
 }	
 
 
-void ExtendedKalmanFilter::correct(Eigen::Matrix6f* xe, Eigen::Matrix6f* pe){
+void ExtendedKalmanFilter::Correct(){
 	
-	Eigen::Matrix6f P;
-	Eigen::Vector6f xp;
-	predict( &xp,  &P);
+      Eigen::VectorXf Meas(6);
+
+      Meas<< odometry_private_.position[0],
+	     odometry_private_.position[1],
+	     odometry_private_.position[2],
+	     odometry_private_.velocity[0],
+	     odometry_private_.velocity[1],
+	     odometry_private_.velocity[2];
+
 	
-	Eigen::Matrix6f Hp;
+       Eigen::MatrixXf K(6,6);
+       K = P_ * Hp_ * (Hp_.transpose() * P_ * Hp_ + Rp_std_).inverse();
 	
-	Hp << 1, 0, 0, 0, 0, 0,
-		  0, 1, 0, 0, 0, 0,
-		  0, 0, 1, 0, 0, 0,
-		  0, 0, 0, 1, 0, 0,
-		  0, 0, 0, 0, 1, 0,
-		  0, 0, 0, 0, 0, 1;
-	
-	Eigen::Matrix6f Rp;
-	
-	Rp << 0.01,    0,    0,    0,    0,    0,
-			 0, 0.01,    0,    0,    0,    0,
-			 0,    0, 0.01,    0,    0,    0,
-			 0,    0,    0, 0.01,    0,    0,
-			 0,    0,    0,    0, 0.01,    0,
-			 0,    0,    0,    0,    0, 0.01;
-	
-	Rp = pow(Rp,2);
-	
-	Eigen::Matrix6f K = P_ * Hp * (Hp.transpose() * P_ * Hp + Rp).inverse();
-	
-	Eigen::Matrix6f *pe = P - K * Hp.transpose() * P_;
-	Eigen::Matrix6f *xe = xp_ + K * (y - Hp * xp_);
+       Pe_ = P_ - K * Hp_.transpose() * P_;
+		
+       Xe_ = Xp_ + K * (Meas - Hp_ * Xp_);
 	
 	
 }
 
-void ExtendedKalmanFilter::CallbackEstimate(const ros::TimerEvent& event){
-     
-     predict(&xp, &P);
-     correct(&xp, &pe);
-}
 
 
 }