+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog for package teamsannio_med_control
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.1.3 (2018-05-25)
+-----------
+* added the data storage section. Such section has been inserted in the position controller node and allows to storage, in defined csv files, the aircraft and controller state.
+* bug fixing
+* the sim time has been added in the launch file. Now the position controller node uses the simulation and not wall clock time.
+* Contributors: Giuseppe Silano, Luigi Iannelli
+
 0.1.2 (2018-05-15)
 -----------
 * added the basic.world file inside the repository